backend/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
python-dotenv==1.2.2
httpx==0.28.1
pydantic>=2.7.0
groq